move clone logic to its own function
