+++ action/action.go
@@ -14,10 +14,10 @@ import (
 	"github.com/observiq/bindplane-op-action/internal/client/config"
 	"github.com/observiq/bindplane-op-action/internal/client/model"
 	"github.com/observiq/bindplane-op-action/internal/client/version"
+	"github.com/observiq/bindplane-op-action/internal/repo"
 	"gopkg.in/yaml.v3"
 
 	"github.com/go-git/go-git/v5"
-	"github.com/go-git/go-git/v5/plumbing"
 	"github.com/go-git/go-git/v5/plumbing/object"
 	"go.uber.org/zap"
 )
@@ -365,35 +365,11 @@ func (a *Action) AutoRollout() error {
 }
 
 func (a *Action) WriteBack() error {
-	// TODO(jsirianni): githubURL should be assembled outside of this package
-	// and passed in. We could remove the need for token, actor, and repo.
-	cloneURL := a.githubURL
-	githubActor := os.Getenv("GITHUB_ACTOR")
-	githubRepo := os.Getenv("GITHUB_REPOSITORY")
-	if cloneURL == "" {
-		cloneURL = fmt.Sprintf(
-			"https://%s:%s@github.com/%s.git",
-			githubActor,
-			a.githubToken,
-			githubRepo,
-		)
-	}
-
 	a.Logger.Info(
 		"Cloning repository", zap.String("branch", a.configurationOutputBranch),
 	)
 
-	// TODO(jsirianni): This context sets the clone timeout. This should be
-	// a configurable option.
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*120)
-	defer cancel()
-
-	repo, err := git.PlainCloneContext(ctx, "./out_repo", false, &git.CloneOptions{
-		URL:           cloneURL,
-		Progress:      os.Stdout,
-		SingleBranch:  true,
-		ReferenceName: plumbing.NewBranchReferenceName(a.configurationOutputBranch),
-	})
+	repo, err := repo.CloneRepo(a.githubURL, a.configurationOutputBranch, a.githubToken)
 	if err != nil {
 		return fmt.Errorf("clone repository: %w", err)
 	}
